Resolving circular import
